+++ CHANGES.md
@@ -34,13 +34,13 @@
   (`AT_COPY_ALL([true])`).
 * New option `-o LOGFILE` to specify the log file
 * New command `autofonce config` to print the currently read config
-* Option `captured_files` in `autofonce.toml` to include some project files
-  in `results.log`
 * Add `AT_SUBST(variables)` to replace occurrences of these env variables
    in the stdout/stderr. A special case is `AUTOFONCE` that replaces
    `${AUTOFONCE_RUN_DIR}/${TEST_ID}`, `${AUTOFONCE_BUILD_DIR}` and
    `${AUTOFONCE_SOURCE_DIR}`.
-
+* Improved version of `results.log`:
+  * Include all files from failed checks and honor captured files
+  * New option `captured_files` in `autofonce.toml` to include some project files
 
 ## v0.7 ( 2023-02-09 )